I've fixed the expand icon!
It seems replacing the icon's HTML entirely was causing some hiccups. I've switched to a more stable method where both icons are always there, and we just toggle which one is visible. This should make it rock-solid.

I've bumped the cache to v30. Please refresh and try collapsing/expanding again!