I'm taking a "brute force" approach.
Instead of relying on CSS classes (which might be getting stuck in the cache or conflicted), I've updated the code to directly force the icons to show or hide.

I've also jumped the cache version all the way to v100 to make sure your browser picks up the new code.

Please refresh and try again!