I found the conflict!
My previous "brute force" fix was fighting with a "hidden" rule that had an !important tag on it. The !important tag was winning, keeping the icon hidden even when I told it to show.

I've updated the code to remove that "hidden" rule and force the icon to show. This double-tap approach guarantees it will be visible.

I've bumped the cache to v101. Please refresh and try again!